+++ src/devices/heatpump.cpp
@@ -29,25 +29,126 @@ Heatpump::Heatpump(uint8_t device_type, uint8_t device_id, uint8_t product_id, c
     LOG_DEBUG(F("Adding new Heat Pump module with device ID 0x%02X"), device_id);
 
     // telegram handlers
-    register_telegram_type(0x047B, F("HP1"), true, [&](std::shared_ptr<const Telegram> t) { process_HPMonitor1(t); });
-    register_telegram_type(0x042B, F("HP2"), true, [&](std::shared_ptr<const Telegram> t) { process_HPMonitor2(t); });
+    register_telegram_type(0x042B, F("HP1"), true, [&](std::shared_ptr<const Telegram> t) { process_HPMonitor1(t); });
+    register_telegram_type(0x047B, F("HP2"), true, [&](std::shared_ptr<const Telegram> t) { process_HPMonitor2(t); });
+
+    // API call
+    Command::add_with_json(this->device_type(), F("info"), [&](const char * value, const int8_t id, JsonObject & object) {
+        return command_info(value, id, object);
+    });
+}
+
+bool Heatpump::command_info(const char * value, const int8_t id, JsonObject & output) {
+    return (export_values(output));
+}
+
+// creates JSON doc from values
+// returns false if empty
+bool Heatpump::export_values(JsonObject & output) {
+    if (Helpers::hasValue(airHumidity_)) {
+        output["airHumidity"] = (float)airHumidity_ / 2;
+    }
+
+    if (Helpers::hasValue(dewTemperature_)) {
+        output["dewTemperature"] = dewTemperature_;
+    }
+
+    return output.size();
 }
 
 void Heatpump::device_info_web(JsonArray & root) {
+    // fetch the values into a JSON document
+    StaticJsonDocument<EMSESP_MAX_JSON_SIZE_MEDIUM> doc;
+    JsonObject                                      output = doc.to<JsonObject>();
+    if (!export_values(output)) {
+        return; // empty
+    }
+
+    print_value_json(root, F("airHumidity"), nullptr, F_(airHumidity), F_(percent), output);
+    print_value_json(root, F("dewTemperature"), nullptr, F_(dewTemperature), F_(degrees), output);
 }
 
 // display all values into the shell console
 void Heatpump::show_values(uuid::console::Shell & shell) {
-    // EMSdevice::show_values(shell); // always call this to show header
+    EMSdevice::show_values(shell); // always call this to show header
+
+    // fetch the values into a JSON document
+    StaticJsonDocument<EMSESP_MAX_JSON_SIZE_MEDIUM> doc;
+    JsonObject                                      output = doc.to<JsonObject>();
+    if (!export_values(output)) {
+        return; // empty
+    }
+
+    print_value_json(shell, F("airHumidity"), nullptr, F_(airHumidity), F_(percent), output);
+    print_value_json(shell, F("dewTemperature"), nullptr, F_(dewTemperature), F_(degrees), output);
 }
 
 // publish values via MQTT
 void Heatpump::publish_values(JsonObject & data, bool force) {
+    // handle HA first
+    if (Mqtt::mqtt_format() == Mqtt::Format::HA) {
+        register_mqtt_ha_config(force);
+    }
+
+    StaticJsonDocument<EMSESP_MAX_JSON_SIZE_MEDIUM> doc;
+    JsonObject                                      output = doc.to<JsonObject>();
+    if (export_values(output)) {
+        Mqtt::publish(F("heatpump_data"), doc.as<JsonObject>());
+    }
+}
+
+void Heatpump::register_mqtt_ha_config(bool force) {
+    if ((mqtt_ha_config_ && !force)) {
+        return;
+    }
+
+    if (!Mqtt::connected()) {
+        return;
+    }
+
+    // Create the Master device
+    StaticJsonDocument<EMSESP_MAX_JSON_SIZE_MEDIUM> doc;
+    doc["name"]    = F_(EMSESP);
+    doc["uniq_id"] = F_(heatpump);
+    doc["ic"]      = F_(iconheatpump);
+
+    char stat_t[50];
+    snprintf_P(stat_t, sizeof(stat_t), PSTR("%s/heatpump_data"), System::hostname().c_str());
+    doc["stat_t"] = stat_t;
+
+    doc["val_tpl"] = F("{{value_json.airHumidity}}");
+
+    JsonObject dev = doc.createNestedObject("dev");
+    dev["name"]    = F("EMS-ESP Heat Pump");
+    dev["sw"]      = EMSESP_APP_VERSION;
+    dev["mf"]      = this->brand_to_string();
+    dev["mdl"]     = this->name();
+    JsonArray ids  = dev.createNestedArray("ids");
+    ids.add("ems-esp-heatpump");
+    Mqtt::publish_retain(F("homeassistant/sensor/ems-esp/heatpump/config"), doc.as<JsonObject>(), true); // publish the config payload with retain flag
+
+    Mqtt::register_mqtt_ha_sensor(nullptr, nullptr, F_(airHumidity), this->device_type(), "airHumidity", F_(percent), nullptr);
+    Mqtt::register_mqtt_ha_sensor(nullptr, nullptr, F_(dewTemperature), this->device_type(), "dewTemperature", F_(degrees), nullptr);
+
+    mqtt_ha_config_ = true; // done
 }
 
 // check to see if values have been updated
 bool Heatpump::updated_values() {
+    if (changed_) {
+        changed_ = false;
+        return true;
+    }
     return false;
+}
+
+/*
+ * Type 0x47B - HeatPump Monitor 2
+ * e.g. "38 10 FF 00 03 7B 08 24 00 4B"
+ */
+void Heatpump::process_HPMonitor2(std::shared_ptr<const Telegram> telegram) {
+    changed_ |= telegram->read_value(dewTemperature_, 0);
+    changed_ |= telegram->read_value(airHumidity_, 1);
 }
 
 #pragma GCC diagnostic push
@@ -58,14 +159,6 @@ bool Heatpump::updated_values() {
  * e.g. "38 10 FF 00 03 2B 00 D1 08 2A 01"
  */
 void Heatpump::process_HPMonitor1(std::shared_ptr<const Telegram> telegram) {
-    // still to implement
-}
-
-/*
- * Type 0x47B - HeatPump Monitor 2
- * e.g. "38 10 FF 00 03 7B 08 24 00 4B"
- */
-void Heatpump::process_HPMonitor2(std::shared_ptr<const Telegram> telegram) {
     // still to implement
 }